+++ NEWS.md
@@ -25,3 +25,4 @@
 # Racmacs 1.1.12
 * Fix error in Racmacs when calculating antigen and sera mean stress per titer
 * Fix error in viewer code when accounting for antigen reactivity adjustments in stress calculations
+* Add option in `plot.acmap()` to draw stress lines